requirements.txt
pandas==2.2.0
numpy==1.24.3
requests==2.31.0
aiohttp==3.9.3
ccxt==4.1.78

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.1
schedule==1.2.1
# Scheduling
APScheduler==3.10.4

# Blockchain
alchemy-sdk==2.10.1

